adding admin directory, fixing style detiails like scrollbar and colors in mainview and unificate the preloads need more work to finish orders viewq
