+++ src/admin/m.js
@@ -1,0 +1,9 @@
+
+function sioji() {
+  return {
+    'hola': 123
+  }
+}
+
+
+module.exports = sioji;
