+++ src/admin/n.js
@@ -1,0 +1,5 @@
+
+const m = require('./m.js');
+
+
+console.log(m())
