+++ src/cajero/orders/orders.html
@@ -8,11 +8,22 @@
 	<meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
 	<meta http-equiv="X-UA-Compatible" content="ie=edge">
 	<!-- Style -->
-	<link href="orders.css" rel="stylesheet" >
+	<link rel="stylesheet" href="orders.css">
+	<link rel="stylesheet" href="../shared/index.css">
 	<title>El Tiburon | Registro De Pedidos</title>
 </head>
 <body>
-	<div class="all-content">
+	<div class="navigator">
+		<div class="navigator-body">
+			<a class="nav-link" href="../tellerView/index.html">
+				<p>Cobro</p>
+			</a>
+			<a class="nav-link marked" href="../orders/orders.html">
+				<p>Pedidos</p>
+			</a>
+		</div>
+	</div>
+	<main class="all-content">
 
 		<div class="table-main">
 			<table class="table">
@@ -36,7 +47,8 @@
 			</div>
 		</div>
 
-	</div>
+	</main>
 </body>
 	<script src="orders.js"></script>
+	<script src="../shared/body.js"></script>
 </html>
