+++ src/cajero/shared/body.js
@@ -1,0 +1,5 @@
+
+
+window.addEventListener('load', () => {
+  document.getElementsByTagName('main')[0].style.opacity = 1;
+})
